+++ CHANGELOG.md
@@ -3,5 +3,9 @@ All notable changes to the "vscode-js-import" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.3.1] - TBD
+- support as in import statement (currently not support line feed)
+- clean up import although is already import
+
+## [0.2.1] - 2017-05-26
 - Initial release